requirements.txt
# Core dependencies
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0

# LiveAvatar dependencies
diffusers>=0.30.0
transformers>=4.44.0
accelerate>=0.33.0
safetensors>=0.4.0
einops>=0.8.0
omegaconf>=2.3.0

# Audio processing
elevenlabs>=1.0.0
pydub>=0.25.0
librosa>=0.10.0
soundfile>=0.12.0

# Video processing
imageio>=2.34.0
imageio-ffmpeg>=0.4.9
opencv-python>=4.9.0
pillow>=10.0.0

# API clients
openai>=1.0.0
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.0
numpy>=1.26.0
huggingface-hub>=0.23.0

# Optional: Flash Attention (install separately)
# pip install flash-attn --no-build-isolation

# Optional: OBS integration
# obsws-python>=1.6.0
